+++ src/controllers/WorkerController.js
@@ -1,10 +1,17 @@
 const Worker = require("../models/Worker");
 const bcrypt = require("bcrypt");
+const workerValidation = require("../validations/workerValidation");
 
 class WorkerController {
   async getWorkerById(req, res) {
     try {
       const { id } = req.params;
+      if (isNaN(id)) {
+        return res
+          .status(400)
+          .json({ error: "El ID debe ser un número válido." });
+      }
+
       const worker = await Worker.findByPk(id);
       if (!worker) {
         return res.status(404).json({ error: "Trabajador no encontrado" });
@@ -17,8 +24,9 @@ class WorkerController {
 
   async createWorker(req, res) {
     try {
-      const { name, email, avatar, password } = req.body;
+      await workerValidation.validate(req.body, { abortEarly: false });
 
+      const { name, email, avatar, password } = req.body;
       const hashedPassword = await bcrypt.hash(password, 10);
 
       const worker = await Worker.create({
@@ -29,6 +37,17 @@ class WorkerController {
       });
       res.status(201).json(worker);
     } catch (error) {
+      if (error.name === "ValidationError") {
+        return res.status(400).json({
+          message: "Errores de validación en los datos proporcionados.",
+          detalles: error.errors,
+        });
+      }
+      if (error.name === "SequelizeUniqueConstraintError") {
+        return res
+          .status(400)
+          .json({ error: "El correo electrónico ya está en uso." });
+      }
       res.status(500).json({ error: "Error al crear el trabajador" });
     }
   }
@@ -36,12 +55,27 @@ class WorkerController {
   async updateWorker(req, res) {
     try {
       const { id } = req.params;
+      if (isNaN(id)) {
+        return res
+          .status(400)
+          .json({ error: "El ID debe ser un número válido." });
+      }
+
+      await workerValidation.validate(req.body, { abortEarly: false });
+
       const updated = await Worker.update(req.body, { where: { id } });
       if (!updated[0]) {
         return res.status(404).json({ error: "Trabajador no encontrado" });
       }
+
       res.json({ message: "Trabajador actualizado correctamente" });
     } catch (error) {
+      if (error.name === "ValidationError") {
+        return res.status(400).json({
+          message: "Errores de validación en los datos proporcionados.",
+          detalles: error.errors,
+        });
+      }
       res.status(500).json({ error: "Error al actualizar el trabajador" });
     }
   }
@@ -49,10 +83,17 @@ class WorkerController {
   async deleteWorker(req, res) {
     try {
       const { id } = req.params;
+      if (isNaN(id)) {
+        return res
+          .status(400)
+          .json({ error: "El ID debe ser un número válido." });
+      }
+
       const deleted = await Worker.destroy({ where: { id } });
       if (!deleted) {
         return res.status(404).json({ error: "Trabajador no encontrado" });
       }
+
       res.json({ message: "Trabajador eliminado correctamente" });
     } catch (error) {
       res.status(500).json({ error: "Error al eliminar el trabajador" });
